fix: 修复插件元数据缺失时不报错的问题
